specify build with hatch to create module
Create the module "whatsmybitrate" so that the program can be called via

`python -m whatsmybitrate ...` without any explicit reference to the full path.

Build is accomplished by the more standard

`pip install -e .`

In the parent commit, to run this script from directories other than the build
directory, one would have to call

`python <path-to>/whatsmybitrate ...`, which is a bit cumbersome.